Update BuildStyle to use float for laser focus

+++ App/Model.cpp
@@ -31,13 +31,13 @@ BuildStyle::~BuildStyle()
 }
 
 void BuildStyle::setStyle(uint64_t bid,
-              uint64_t focus,
-              float power,
-              uint64_t pExpTime,
-              uint64_t pDistTime,
-              float speed,
-              uint64_t lId,
-              LaserMode lMode)
+                          float focus,
+                          float power,
+                          uint64_t pExpTime,
+                          uint64_t pDistTime,
+                          float speed,
+                          uint64_t lId,
+                          LaserMode lMode)
 {
     // Convenience function
     id = bid;
